+++ crates/stackable-operator/CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 - BREAKING: Add a new CLI flag/env to disabling CRD maintenance: `--disable-crd-maintenance` ([#1085]).
 
+### Changed
+
+- BREAKING: Upgrade to `schemars` 1.0, `kube` 2.0 and `k8s-openapi` 0.26 (using Kubernetes 1.34) ([#1091]).
+
 ### Removed
 
 - BREAKING: Remove the Merge implementation for PodTemplateSpec ([#1087]).
@@ -23,6 +27,7 @@ All notable changes to this project will be documented in this file.
 [#1085]: https://github.com/stackabletech/operator-rs/pull/1085
 [#1087]: https://github.com/stackabletech/operator-rs/pull/1087
 [#1090]: https://github.com/stackabletech/operator-rs/pull/1090
+[#1091]: https://github.com/stackabletech/operator-rs/pull/1091
 
 ## [0.96.0] - 2025-08-25